PageIndex/requirements.txt
pymupdf==1.24.11
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
tiktoken>=0.7.0
